requirements.txt
streamlit==1.41.1
openai==1.58.1
python-dotenv==1.0.1
requests==2.32.3
attackcti==0.4.1
pandas==2.2.3
plotly==5.24.1
shodan==1.31.0
reportlab==4.2.2